add admin product management and tests

+++ app/schemas/product.py
@@ -36,6 +36,6 @@ class ProductResponse(BaseModel):
 class ProductUpdate(BaseModel):
     name: Optional[str] = None
     description: Optional[str] = None
-    price: Optional[float] = None
+    price: Optional[Decimal] = None
     stock: Optional[int] = None
     category_id: Optional[int] = None
